update integration test module structure for new route structure
