+++ auth-service/tests/api/logout.rs
@@ -1,0 +1,8 @@
+use crate::helpers::TestApp;
+
+#[tokio::test]
+async fn logout_route_works() {
+    let app = TestApp::new().await;
+    let response = app.post_logout().await;
+    assert_eq!(response.status().as_u16(), 200);
+}
