+++ auth-service/tests/api/root.rs
@@ -1,0 +1,11 @@
+use crate::helpers::TestApp;
+
+#[tokio::test]
+async fn root_returns_auth_ui() {
+    let app = TestApp::new().await;
+
+    let response = app.get_root().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+    assert_eq!(response.headers().get("content-type").unwrap(), "text/html");
+}
